pyproject.toml
[tool.poetry]
name = "viscos"
version = "0.1.0"
description = ""
authors = ["Nils Ramolla <nilsramolla@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"
numpy = "^2.2.1"
opencv-python = "^4.10.0.84"
scikit-learn = "^1.6.0"
scikit-image = "^0.25.0"
plotly = "^5.24.1"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F401","D102", "D103", "E225", "E226", "I001"]